requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
djangorestframework==3.13.1
executing==0.8.3
Flask==3.0.3
gunicorn==20.1.0
idna==3.8
ipdb==0.13.9
ipython==8.1.1
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.3
MouseInfo==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pywhatkit==5.4
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
urllib3==2.2.2
uuid==1.30
wcwidth==0.2.5
Werkzeug==3.0.4
wikipedia==1.4.0
